% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbleplot-package.R
\docType{package}
\name{bubbleplot-package}
\alias{bubbleplot-package}
\title{Bubble Plot, a Scatter Plot with Varying Symbol Sizes and Colors}
\description{
Draw a bubble plot, a scatterplot with varying symbol sizes and colors, or
add points to existing plots. A variety of input formats are supported,
including vectors, matrices, data frames, formulas, etc.
}
\details{
\emph{Plot:}
\tabular{ll}{
  \code{\link{bubbleplot}} \tab Bubble plot\cr
}
}
\author{
Arni Magnusson.
}
