% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbleplot.R
\name{bubbleplot}
\alias{bubbleplot}
\alias{bubbleplot.default}
\alias{bubbleplot.formula}
\title{Bubble Plot}
\usage{
bubbleplot(x, ...)

\method{bubbleplot}{default}(x, y, z, std = TRUE, pow = 0.5, add = FALSE,
  rev = FALSE, type = "p", ylim = NULL, xlab = NULL, ylab = NULL,
  pch = c(16, 1), cex.points = 1, col = "black", bg = par("bg"), ...)

\method{bubbleplot}{formula}(formula, data, subset, na.action = NULL, ...)
}
\arguments{
\item{x}{a vector of values for the horizontal axis. Can also be a
2-dimensional matrix or table (x values in column names and y values
in rows), or a data frame containing \code{x}, \code{y}, and \code{z}
in that order.}

\item{\dots}{passed to \code{bubbleplot.*} methods.}

\item{y}{a vector of values for the vertical axis.}

\item{z}{a vector of values determining the bubble sizes.}

\item{std}{whether to standardize \code{z} by dividing with \code{mean(z)}.}

\item{pow}{a power coefficient for the bubble sizes (\code{0.5} is the square
root).}

\item{add}{whether to add bubbles to an existing plot.}

\item{rev}{whether to reverse the y axis.}

\item{type}{passed to \code{points}.}

\item{ylim}{passed to \code{plot}.}

\item{xlab}{passed to \code{plot}.}

\item{ylab}{passed to \code{plot}.}

\item{pch}{passed to \code{points}.}

\item{cex.points}{scales all bubble sizes.}

\item{col}{passed to \code{points}.}

\item{bg}{passed to \code{points}.}

\item{formula}{has the form \code{z ~ x + y}, where \code{z} determines the
bubble sizes and \code{x} and \code{y} determine bubble locations.}

\item{data}{where formula terms are stored, e.g. data frame or list.}

\item{subset}{a logical vector specifying which data to plot.}

\item{na.action}{how \code{NA} values are handled.}
}
\description{
Draw a bubble plot, a scatterplot with varying symbol sizes and colors.
}
\note{
The \code{pch}, \code{col}, and \code{bg} arguments can be be vectors of
length 2, where negative \code{z} values are drawn in \code{pch[2]},
\code{col[2]}, and \code{bg[2]}.
}
\examples{
# Tree circumference
bubbleplot(circumference~Tree+age, Orange, pow=1, cex=2)

library(nlme)
fm <- nlme(circumference~phi1/(1+exp(-(age-phi2)/phi3)),
           fixed=phi1+phi2+phi3~1, random=phi1~1|Tree,
           data=Orange, start=c(phi1=200,phi2=800,phi3=400))

# Model residuals
bubbleplot(residuals(fm)~Tree+age, Orange)
bubbleplot(residuals(fm)~Tree+age, Orange, cex=2.5, pch=16,
           col=c("dodgerblue","orange"))

# Richter magnitude, amplitude, and energy release
bubbleplot(mag~long+lat, quakes, pch=1)
bubbleplot(10^mag~long+lat, quakes, cex=1.2, col=gray(0, 0.3))
bubbleplot(sqrt(1000)^mag~long+lat, quakes, cex=1.2, col=gray(0, 0.3))

}
\seealso{
\code{\link{points}} is the underlying function used to draw the bubbles.

\code{\link{symbols}} can also draw bubbles, but does not handle negative
\code{z} values or have convenience features such as \code{pow} and
\code{rev}.

\code{\link{bubbleplot-package}} gives an overview of the package.
}
