% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbleplot-package.R
\docType{package}
\name{bubbleplot-package}
\alias{bubbleplot-package}
\title{Bubble Plot, a Scatter Plot with Varying Symbol Sizes and Colors}
\description{
Draw a bubble plot, a scatterplot with varying symbol sizes and colors, or
add points to existing plots. A variety of input formats are supported,
including vectors, matrices, data frames, formulas, etc.
}
\details{
\emph{Plot:}
\tabular{ll}{
  \code{\link{bubbleplot}} \tab Bubble plot\cr
}
\emph{Examples:}
\tabular{ll}{
  \code{\link{catch.d}} \tab Catch at age data frame\cr
  \code{\link{catch.m}} \tab Catch at age matrix\cr
  \code{\link{catch.r}} \tab Catch at age residuals
}
}
\author{
Arni Magnusson.
}
